+++ Trainings/src/main/java/com/iiht/mentor/trainings/model/User.java
@@ -28,7 +28,7 @@ import org.hibernate.annotations.NaturalId;
 public class User{
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String name;
 
@@ -59,15 +59,15 @@ public class User{
         this.confirmCode = confirmCode;
     }
 
-    public Long getId() {
-        return id;
-    }
+    public int getId() {
+		return id;
+	}
 
-    public void setId(Long id) {
-        this.id = id;
-    }
+	public void setId(int id) {
+		this.id = id;
+	}
 
-    public String getUsername() {
+	public String getUsername() {
         return username;
     }
