+++ UserProfile/src/main/java/com/iiht/mentor/userprofile/model/UserProfileEntity.java
@@ -25,7 +25,7 @@ public class UserProfileEntity  {
 	
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @NotBlank
     @Size(min=3, max = 50)
@@ -66,15 +66,15 @@ public class UserProfileEntity  {
         this.confirmCode = confirmCode;
     }
 
-    public Long getId() {
-        return id;
-    }
+    public int getId() {
+		return id;
+	}
 
-    public void setId(Long id) {
-        this.id = id;
-    }
+	public void setId(int id) {
+		this.id = id;
+	}
 
-    public String getUsername() {
+	public String getUsername() {
         return username;
     }
